Refactor jattach such that it can be structured as library, i.e. for use in other open source projects. Essentially, we just pull main() into its own file.
